Remove plutus-chain-index-core dep
