+++ cardano-node-socket-emulator/src/Cardano/Node/Socket/Emulator/Types.hs
@@ -19,27 +19,91 @@
 module Cardano.Node.Socket.Emulator.Types where
 
 import Cardano.Api (NetworkId, Value)
+import Cardano.Chain.Slotting (EpochSlots (..))
+import Cardano.Ledger.Block qualified as CL
+import Cardano.Ledger.Era qualified as CL
+import Cardano.Ledger.Shelley.API (extractTx, unsafeMakeValidated)
 import Cardano.Node.Emulator.Internal.Node (ChainControlEffect, ChainEffect, ChainEvent, SlotConfig, fromBlockchain,
                                             testnet, unsafeMakeValid)
-import Cardano.Node.Socket.Emulator.Chain (MockNodeServerChainState, fromEmulatorChainState)
+import Cardano.Node.Emulator.Internal.Node.Chain qualified as EC
+import Codec.Serialise (DeserialiseFailure)
+import Codec.Serialise qualified as CBOR
+import Control.Concurrent.STM
 import Control.Lens (makeLenses)
+import Control.Monad (forever)
+import Control.Monad.Class.MonadST (MonadST)
+import Control.Monad.Class.MonadTimer (MonadDelay (threadDelay), MonadTimer)
 import Control.Monad.Freer.Extras.Log (LogMessage, LogMsg)
 import Control.Monad.Freer.State qualified as Eff
-import Control.Monad.IO.Class (MonadIO)
+import Control.Monad.IO.Class (MonadIO, liftIO)
+import Crypto.Hash (SHA256, hash)
 import Data.Aeson (FromJSON, ToJSON)
+import Data.Aeson.Extras qualified as JSON
+import Data.ByteArray qualified as BA
+import Data.ByteString.Lazy qualified as BSL
+import Data.ByteString.Short qualified as BS
+import Data.Coerce (coerce)
 import Data.Default (Default, def)
+import Data.Foldable (toList)
+import Data.Functor (void)
 import Data.Map qualified as Map
+import Data.Maybe (listToMaybe)
+import Data.Sequence.Strict (fromList)
+import Data.Text qualified as Text
 import Data.Time.Clock (UTCTime)
 import Data.Time.Format.ISO8601 qualified as F
 import Data.Time.Units (Millisecond)
 import Data.Time.Units.Extra ()
+import Data.Void (Void)
 import GHC.Generics (Generic)
+import Ledger (Block, CardanoTx, OnChainTx (..), Slot (..))
 import Ledger.Address (CardanoAddress)
 import Ledger.CardanoWallet
-import Ledger.Index (createGenesisTransaction)
+import Ledger.Index (UtxoIndex, createGenesisTransaction)
+import Ledger.Test (testNetworkMagic)
+import Network.TypedProtocol.Codec (Codec)
+import Ouroboros.Consensus.Byron.Ledger qualified as Byron
+import Ouroboros.Consensus.Cardano.Block (CardanoBlock, CodecConfig (..))
+import Ouroboros.Consensus.Cardano.Block qualified as OC
+import Ouroboros.Consensus.HardFork.Combinator.AcrossEras (OneEraHash (..))
+import Ouroboros.Consensus.Ledger.SupportsMempool (ApplyTxErr)
+import Ouroboros.Consensus.Network.NodeToClient (ClientCodecs, cChainSyncCodec, cTxSubmissionCodec, clientCodecs)
+import Ouroboros.Consensus.Node.NetworkProtocolVersion (BlockNodeToClientVersion, supportedNodeToClientVersions)
+import Ouroboros.Consensus.Protocol.Praos.Header qualified as Praos
+import Ouroboros.Consensus.Shelley.Eras (StandardCrypto)
+import Ouroboros.Consensus.Shelley.Ledger qualified as Shelley
+import Ouroboros.Network.Block (Point)
+import Ouroboros.Network.Block qualified as Ouroboros
+import Ouroboros.Network.Mux
+import Ouroboros.Network.NodeToClient (NodeToClientVersion (..), NodeToClientVersionData (..))
+import Ouroboros.Network.Protocol.ChainSync.Type qualified as ChainSync
+import Ouroboros.Network.Protocol.LocalTxSubmission.Type qualified as TxSubmission
+import Ouroboros.Network.Util.ShowProxy
 import Prettyprinter (Pretty, pretty, viaShow, vsep, (<+>))
+import Prettyprinter.Extras (PrettyShow (PrettyShow))
 import Servant.Client (BaseUrl (BaseUrl, baseUrlPort), Scheme (Http))
 
+type Tip = Ouroboros.Tip (CardanoBlock StandardCrypto)
+
+type TxPool = [CardanoTx]
+
+data MockNodeServerChainState = MockNodeServerChainState
+  { _txPool      :: TxPool
+  , _index       :: UtxoIndex
+  , _currentSlot :: Slot
+  , _channel     :: TChan Block
+  , _tip         :: Tip
+  } deriving (Generic)
+
+makeLenses ''MockNodeServerChainState
+
+instance Show MockNodeServerChainState where
+    -- Skip showing the full chain
+    show MockNodeServerChainState {_txPool, _index, _currentSlot, _tip} =
+        "MockNodeServerChainState { " <> show _txPool
+                        <> ", " <> show _index
+                        <> ", " <> show _currentSlot
+                        <> ", " <> show _tip <> " }"
 
 -- | Node server configuration
 data NodeServerConfig =
@@ -116,6 +180,21 @@ data AppState =
 
 makeLenses 'AppState
 
+-- | Build a CNSE ChainState from a emulator ChainState
+fromEmulatorChainState :: MonadIO m => EC.ChainState -> m MockNodeServerChainState
+fromEmulatorChainState EC.ChainState {EC._txPool, EC._index, EC._chainCurrentSlot, EC._chainNewestFirst} = do
+    ch <- liftIO $ atomically newTChan
+    void $ liftIO $
+        mapM_ (atomically . writeTChan ch) _chainNewestFirst
+    pure $ MockNodeServerChainState { _channel     = ch
+                      , _txPool      = _txPool
+                      , _index       = _index
+                      , _currentSlot = _chainCurrentSlot
+                      , _tip         = case listToMaybe _chainNewestFirst of
+                                        Nothing -> Ouroboros.TipGenesis
+                                        Just block -> Ouroboros.Tip (fromIntegral _chainCurrentSlot) (coerce $ blockId block) (fromIntegral _chainCurrentSlot)
+                      }
+
 -- | 'ChainState' with initial values
 initialChainState :: MonadIO m => Map.Map CardanoAddress Value -> m MockNodeServerChainState
 initialChainState =
@@ -140,3 +219,96 @@ instance Pretty CNSEServerLogMsg where
             <+> "Slot length:" <+> viaShow slotLength
         StartingCNSEServer p   -> "Starting Cardano Node Emulator on port" <+> pretty p
         ProcessingChainEvent e -> "Processing chain event" <+> pretty e
+
+-- | The node protocols require a block header type.
+newtype BlockId = BlockId { getBlockId :: BS.ShortByteString }
+  deriving (Eq, Ord, Generic)
+  deriving newtype (CBOR.Serialise)
+  deriving Pretty via (PrettyShow BlockId)
+
+instance Show BlockId where
+    show = Text.unpack . JSON.encodeByteString . BS.fromShort . getBlockId
+
+-- | A hash of the block's contents.
+blockId :: Block -> BlockId
+blockId = BlockId
+        . BS.toShort
+        . BA.convert
+        . hash @_ @SHA256
+        . BSL.toStrict
+        . CBOR.serialise
+
+-- | Protocol versions
+nodeToClientVersion :: NodeToClientVersion
+nodeToClientVersion = NodeToClientV_13
+
+-- | A temporary definition of the protocol version. This will be moved as an
+-- argument to the client connection function in a future PR (the network magic
+-- number matches the one in the test net created by scripts)
+nodeToClientVersionData :: NodeToClientVersionData
+nodeToClientVersionData = NodeToClientVersionData { networkMagic = testNetworkMagic }
+
+-- | A protocol client that will never leave the initial state.
+doNothingInitiatorProtocol
+  :: MonadTimer m => RunMiniProtocol 'InitiatorMode BSL.ByteString m a Void
+doNothingInitiatorProtocol =
+    InitiatorProtocolOnly $ MuxPeerRaw $
+    const $ forever $ threadDelay 1e6
+
+doNothingResponderProtocol
+  :: MonadTimer m => RunMiniProtocol 'ResponderMode BSL.ByteString m Void a
+doNothingResponderProtocol =
+  ResponderProtocolOnly $ MuxPeerRaw $
+  const $ forever $ threadDelay 1e6
+
+-- | Boilerplate codecs used for protocol serialisation.
+
+-- | The number of epochSlots is specific to each blockchain instance. This value
+-- is what the cardano main and testnet uses. Only applies to the Byron era.
+epochSlots :: EpochSlots
+epochSlots = EpochSlots 21600
+
+codecVersion :: BlockNodeToClientVersion (CardanoBlock StandardCrypto)
+codecVersion = versionMap Map.! nodeToClientVersion
+  where
+    versionMap =
+      supportedNodeToClientVersions
+        (Proxy @(CardanoBlock StandardCrypto))
+
+codecConfig :: CodecConfig (CardanoBlock StandardCrypto)
+codecConfig =
+  CardanoCodecConfig
+    (Byron.ByronCodecConfig epochSlots)
+    Shelley.ShelleyCodecConfig
+    Shelley.ShelleyCodecConfig
+    Shelley.ShelleyCodecConfig
+    Shelley.ShelleyCodecConfig
+    Shelley.ShelleyCodecConfig
+
+nodeToClientCodecs
+  :: forall m. MonadST m
+  => ClientCodecs (CardanoBlock StandardCrypto) m
+nodeToClientCodecs =
+  clientCodecs codecConfig codecVersion nodeToClientVersion
+
+-- | These codecs are currently used in the mock nodes and will
+--   probably soon get removed as the mock nodes are phased out.
+chainSyncCodec
+  :: (block ~ CardanoBlock StandardCrypto)
+  => Codec (ChainSync.ChainSync block (Point block) Tip)
+           DeserialiseFailure
+           IO BSL.ByteString
+chainSyncCodec = cChainSyncCodec nodeToClientCodecs
+
+txSubmissionCodec
+  :: (block ~ CardanoBlock StandardCrypto)
+  => Codec (TxSubmission.LocalTxSubmission (Shelley.GenTx block) (ApplyTxErr block))
+           DeserialiseFailure IO BSL.ByteString
+txSubmissionCodec = cTxSubmissionCodec nodeToClientCodecs
+
+toCardanoBlock :: Praos.Header StandardCrypto -> Block -> CardanoBlock StandardCrypto
+toCardanoBlock header block = OC.BlockBabbage (Shelley.mkShelleyBlock $ CL.Block header $ CL.toTxSeq $ fromList $ extractTx . getOnChainTx <$> block)
+
+fromCardanoBlock :: CardanoBlock StandardCrypto -> Block
+fromCardanoBlock (OC.BlockBabbage (Shelley.ShelleyBlock (CL.Block _ txSeq) _)) = map (OnChainTx . unsafeMakeValidated) . toList $ CL.fromTxSeq txSeq
+fromCardanoBlock _ = []
