+++ plutus-chain-index-core/plutus-chain-index-core.cabal
@@ -42,7 +42,6 @@ library
   import:          lang
   exposed-modules:
     Cardano.Protocol.Socket.Client
-    Cardano.Protocol.Socket.Type
     Plutus.ChainIndex
     Plutus.ChainIndex.Api
     Plutus.ChainIndex.ChainIndexError
@@ -65,7 +64,6 @@ library
     Plutus.ChainIndex.Types
     Plutus.ChainIndex.UtxoState
     Plutus.Contract.CardanoAPI
-    Plutus.Monitoring.Util
 
   hs-source-dirs:  src
 
@@ -73,34 +71,25 @@ library
   -- Local components
   --------------------
   build-depends:
-    , cardano-node-emulator  >=1.2.0
-    , freer-extras           >=1.2.0
-    , marconi-chain-index    >=1.2.0
-    , marconi-core           >=1.2.0
-    , plutus-ledger          >=1.2.0
-    , plutus-script-utils    >=1.2.0
+    , cardano-node-emulator         >=1.2.0
+    , cardano-node-socket-emulator  >=1.2.0
+    , freer-extras                  >=1.2.0
+    , marconi-chain-index           >=1.2.0
+    , marconi-core                  >=1.2.0
+    , plutus-ledger                 >=1.2.0
+    , plutus-script-utils           >=1.2.0
 
   --------------------------
   -- Other IOG dependencies
   --------------------------
   build-depends:
-    , cardano-api                   >=1.35
-    , cardano-ledger-byron
-    , cardano-ledger-core
-    , cardano-ledger-shelley
-    , io-classes
+    , cardano-api                  >=1.35
     , iohk-monitoring
-    , ouroboros-consensus
-    , ouroboros-consensus-byron
-    , ouroboros-consensus-cardano
-    , ouroboros-consensus-protocol
-    , ouroboros-consensus-shelley
     , ouroboros-network
     , ouroboros-network-framework
-    , plutus-core                   >=1.0.0
-    , plutus-ledger-api             >=1.0.0
-    , plutus-tx                     >=1.0.0
-    , typed-protocols
+    , plutus-core                  >=1.0.0
+    , plutus-ledger-api            >=1.0.0
+    , plutus-tx                    >=1.0.0
 
   ------------------------
   -- Non-IOG dependencies
@@ -126,7 +115,6 @@ library
     , lens
     , memory
     , mtl
-    , nothunks
     , openapi3
     , prettyprinter              >=1.1.0.1
     , resource-pool
@@ -141,10 +129,7 @@ library
     , servant-swagger-ui
     , sqlite-simple
     , stm
-    , strict-containers
     , text
-    , text-class
-    , unordered-containers
     , warp
 
 test-suite plutus-chain-index-test
