+++ plutus-chain-index/plutus-chain-index.cabal
@@ -56,10 +56,11 @@ library
   -- Local components
   --------------------
   build-depends:
-    , cardano-node-emulator    >=1.2.0
-    , freer-extras             >=1.2.0
-    , plutus-chain-index-core  >=1.2.0
-    , plutus-ledger            >=1.2.0
+    , cardano-node-emulator         >=1.2.0
+    , cardano-node-socket-emulator  >=1.2.0
+    , freer-extras                  >=1.2.0
+    , plutus-chain-index-core       >=1.2.0
+    , plutus-ledger                 >=1.2.0
 
   --------------------------
   -- Other IOG dependencies
